feat(BOUN-944 BOUN-932): switch to new caching library, fix some bugs

* `stretto` crate seems to suck and consume 7GB+ of memory on canary under prod load with nothing in the cache (memory leak or fragmentation). Switch to `moka` crate which I tested in canary and it works nicely, memory usage is less than 700MB on average
* Don't cache non-2xx responses (maybe we should cache 4xx too, not sure, to be discussed)
* Generate two metric gauges with cache stats (number of items in cache and the total size in bytes)
* Improve cache tests a bit
* As a side hustle fix the problem with not logging responses with an empty body - this fixes e.g. 202 responses to `call` and things like this 

See merge request dfinity-lab/public/ic!15899

+++ rs/boundary_node/ic_boundary/src/cli.rs
@@ -158,7 +158,7 @@ pub struct CacheConfig {
     pub cache_size_bytes: Option<u64>,
     /// Maximum size of a single cached response item in bytes
     #[clap(long, default_value = "65536")]
-    pub cache_max_item_size_bytes: usize,
+    pub cache_max_item_size_bytes: u64,
     /// Time-to-live for cache entries in seconds
     #[clap(long, default_value = "1")]
     pub cache_ttl_seconds: u64,
